EAA Adjusting controllers and routes with code inclusion for swagger
